+++ src/pylingdocs/data/format_templates/html/basic/hooks/post_gen_project.py
@@ -14,10 +14,12 @@ for filename in [
     "examples.css",
     "examples.js",
     "glossing.js",
-    "styling.css",
+    "glossing.css",
     "crossref.js",
     "alignment.css",
     "alignment.js",
+    "tables.css",
+    "styling.css",
 ]:
     source = Path(WEB_DIR / filename)
     target = Path(".")